requirements.txt
pyserial>=3.5
matplotlib>=3.8.0
numpy>=1.26.2
pynput>=1.7.6
